Fixed Various Bugs
Fixed various bugs and small errors. Added Map and Tile classes to the Maps namespace. Properly assigned to the tiles array in the Map constructor. Added do-while loop to the ReadFile method in the FileHandler class. Called Debug.Test() from Program.Main(). Altered Draw() method in the Tile class so that the characters are drawn instead of their integerical ascii codes.