Added Entity Collision (Altered Entity, Player, Enemy, Tile, TileProfile & FileHandler Classes)
Removed Entity constructor code and added it to Player and Enemy constructors individually, since position was being assigned to after referencing and occupying a tile. Altered TileProfile so that it reads an integer representing whether or not the tile in question is walkable, when reading a .set file. Altered Tile class to assign to isWalkable. Altered Entity.Move() so that tiles are vacated and occupied appropriately. Altered FileHandler.Setup() to write the correct information to the default.set file.