Added MapFormatter Class
Added MapFormatter to allow maps to be compressed and decompressed using run-length encoding. Altered Map.GetRawData() method to correctly append '\' characters, since one was being appended to the end of the raw data string, thus causing problems. Altered Game constructor so that map files are decompressed or "deserialized" when loading. Fixed a compatibility issue for macOS in Window.ResizeWindow() method. Will change this in a future commit.