Added Display & GUI Classes
Added Display class for writing non-contiguous or column-based text to the console. Added GUI class to improve ease-of-use by providing a visual interface which can be interacted with using keycodes rather than having to type commands via a Console.ReadLine() call. Altered Map.Draw() method so that all rows are drawn on the same starting column or X coordinate of the Console cursor. Altered TileProfile for the same reason. Added a basic menu to the Debug.Test() method.