Implemented Entity Saving (Added Conversion Class, Altered Both Entities and Maps Namespaces)
Added Conversion class to easily allow string literals to be converted to other datatypes (both base and custom types). Added Id property as well as abstract Save() and Load() methods. Added a new constructor to Entity class, overriding it with the same or similar functionality in both Player and Enemy. Altered EntityCollection to add the Array property for reference in MapFormatter methods. Also added EntityCollection.AssignId() method to return a unique id for new entities. Added TileSet property to Map for reference in MapFormatter. Added Name property to TileSet for reference in MapFormatter. Altered MapFormatter.Serialize() and MapFormatter.Deserialize() to allow for the saving of entities, map names and tileset names.