Added SaveManager Class
Added SaveManager class to allow users to save the game they are currently playing to a file and then load it again later. Moved Clear() from Window to Display, since using this scope makes more sense practically. Added SplitStringOverBufferWidth method for aesthetic use in vertical menus. Altered EntityCollection.Add() so that the newly added entity occupies the tile it is standing on. Altered EntityCollection.Clear() to vacate tiles which were previously occupied by any entities in the collection. Altered both constructors and Move() in Player class to call Game.SetCameraPosition(), so that the camera is constantly focused on the player, straight from instantiation. Added GetNumberOfFilesInDirectory() and GetFilesInDirectory() to FileHandler class to allow directories to be inspected more easily. Added some temporary HUD to Game.Start(), so that the player's health can be seen. This was used for debugging saving and loading. Added save and load map controls to the input section of Game.Start(). Added LoadMap() to Game class for easy loading of maps. Added some basic GUI to Game.GameOver(), to make the process of restarting easier. Added YesOrNo() method to GUI class to allow boolean values to be obtained by simply picking 'YES' or 'NO' from a vertical menu. Altered GetKeyDown() and added modifier properties to Input class to allow for the detection of any modifier keys, such as Control, Shift or Alt, being held down during a key press.