Cleaned Up SaveManager & Game Classes
Altered the way in which games are loaded. Changed the return type of SaveManager.Load() from void to Game. Altered the ChooseSlotToSave() and ChooseSlotToLoad() methods in SaveManager, to add more validation and to fix any display bugs which were occuring before. Changed the access modifiers of the static Game.Current property so that public assignments are allowed and, when this happens, the previous value is correctly disposed. Added Frame property to Game, which increments on each iteration of the while-true loop in Game.Start(). Added Game.Stop() to assign the private field isPlaying to false. Added Game.Dispose() to clear any entities remaining in the game's map's entity collection. Altered EntityCollection.Clear() so that all entities' corresponding tiles are vacated, and so that each entity individually falls out of scope. Moved input functionality from Game to Input.GetPlayerInput, seen as this is a more suitable design choice. Changed Map.AREA_SIZE constant from 4 to 20. This value was only at four originally for simplicity. Fixed a bug with the default tileset not being loaded in TileSet.Default property. Moved menu functionality from Debug.Test() to Program.Main(). This is wrapped in a while-true loop, so as soon as a game ends, the menu will be shown.