Improved Map Editor (Altered MapEditor, Placement, Enemy, FileHandler, Game, Input, Map & Program Classes)
Added more validation for operations such as creating a new map, opening another map, or leaving without saving the current one. Altered MapEditor.Open() so that all map files are ordered alphabetically. Added an overload for MapEditor.Open() which takes a map name argument so that a specific map can be opened, rather than displaying a GUI for selecting one. Added MapEditor.EnterPlayMode() and MapEditor.ExitPlayMode(). MapEditor.EnterPlayMode() creates a temporary file containing the map data, and allows the player to immediately playtest the map. MapEditor.ExitPlayMode() returns to the map editor. Added MapEditor.Reset() to reset all necessary fields. Moved map editor debugging functionality from Input.GetEditorInput() to MapEditor.DebuggingMenu(), since it makes more sense that this functionality would be within the MapEditor class. Added validation to Placement.Move() so that the placement point cannot be moved outside the bounds of the map. Added some validation to Placement.Place() so that placing a tile onto a tile of the same type will not allow the method to proceed, saving some processing power. Decreased the value of the argument in Player.Instance.TakeDamage() call from Enemy.Update(), so that the player does not die after moving past an enemy. This functionality will be changed in future commits. Added FileHandler.DeleteFile() so that local temporary files can be deleted. Re-added Test value to GameType enumeration, as the removal of this value was a mistake. Added separate functionality in Game.Start() for the Test state. Made minor changes to the Game.GameOver() method, fixing an issue where the old system for loading games was in use. Added "fast-move" ability in Input.GetEditorInput(), meaning that if the shift key is held down and the placement point is moved, it will move four units along, rather than just one. Added MapEditor.Fill() to allow the user to quickly fill the area containing the placement point with a character, which is given by the user. Added Map.Fill() to allow an area to easily be filled with a single tile. Added an overload of Map.SetTile() which takes a vector and a tile instead of the individual coordinates and the tile. Now, in the main program loop, in Program.Main(), the temporary file created by the MapEditor.EnterPlayMode() method will be deleted if it still exists.