Drastically Changed TileSet Functionality
Due to a major error in my judgement, and a severe lack of concentration, I somehow managed to code the functionality for the encoding of entire tilesets into map files, when the whole point of converting tiles to and from ASCII was to prevent such a need. I have changed the TileSet class into a static class with a static readonly array of tile profiles. Now, all maps will abide by a universal tile set. Assigned hasSaved field a default value of true in MapEditor class, to prevent the 'Are you sure you want to create a new map without saving?' message from being displayed prematurely. Altered FileHandler.Setup() to create a 'custom' directory and a '.hidden' directory inside the maps directory in bin/Debug/netcoreapp3.1/. Altered MapEditor.New(), MapEditor.Save() and MapEditor.Open() to save to and load from the correct directory -- the 'custom' directory. Added missing Input.GetKeyDown() call in MapEditor.Open() when no map files are found. Altered MapEditor.EnterPlayMode() and MapEditor.ExitPlayMode() to save temporary files to and load temporary files from the correct directory -- the '.hidden' directory. Altered tile placement functionality in Placment.Place(), so that for the wall tile in particular, a perspective wall tile will be placed directly beneath the newly placed wall tile. Altered entity placement functionality in Placement.Place() to cause any duplicate player entities to be removed so that only one instance of the player can exist in one map, to avoid errors later on. Altered Game.Start() so that a title is displayed throughout all game modes and added some extra validation to ensure that if there is no player instance in the map, the HUD will not be drawn. Altered MapFormatter.Serialize() and MapFormatter.Deserialize() due to the removal of tileset encoding within map files. Altered existing constructor and added new overload constructor in TileProfile to allow for easy predefining of tiles. Changed window width and height slightly.