Made General Editor Improvements
Altered assignment to position field in Placement.Reset() so that the placement point's location defaults to the position of the player instance, if one is present. Added Display.WriteAtCentreBottom() so that key bindings can be drawn horizontally centred at the bottom of the screen. Added Display.ShowRibbonMessage() for displaying important information, such as a 'Game Over' event. Added Display.DrawMap() so that maps can now be drawn centrally on the console buffer. Removed Placement.Reset() from MapEditor.New() so that the position field can be properly assigned to, once a player instance has been instantiated. Added virtual OnDestroy() method to Entity so that final actions can be performed before the entity is destroyed. Please note this method is NOT a destructor, hence the lack of a tilde character preceeding the identifier. Called Entity.OnDestroy() from EntityCollection.Remove(). Added override of Entity.OnDestroy() in the Player class so that the Instance property is assigned to null to avoid any null reference exceptions occuring later on. Moved removal of temporary map file functionality from Program.Main() to FileHandler.Setup() as this functionality was in the wrong place. Called the new methods in Display class in Game.Start(). Altered GUI.HorizontalBar() so that, when drawn, the bar is aligned with the map. Added a small 'cheat' to Input.GetPlayerInput() and Input.GetTestInput() where the player can deal damage on themselves. This is temporary and is purely for test purposes. Added ToString() override in Tile class, returning the unicode for its profile. Made minor graphical change to one of the tiles in TileSet class. Altered Window.Setup() so that Ctrl+C is treated as input rather than an exit condition, and the correct UTF-8 encoding is now used, which should prevent any visual errors from occuring later.