Added AStar Namespace (Added Grid, Node & Pathfinder Classes)
Added basic pathfinding with the addition of the AStar namespace. Added Grid class to keep track of all nodes in a map. Added Node class to contain pathfinding information such as G, H & F costs. Added Pathfinder class to calculate and return a path, taking a position and a destination. Altered Enemy class to make use of this pathfinding. Removed unnecessary comments from Game.Start(). Added Grid property to Map class. Added Node property and two new constructors to the Tile class to allow for the Node property to be assigned to.

Please note that many of the changes in this commit will be condensed and moved into some already existing classes.