Added Basic Combat (Altered Enemy, Player, Game & Map Classes)
Added basic attacking to Enemy class, so that the player takes damage when too close to an enemy. Added ability to attack in the direction of the player's movement, if they are standing adjacent to an enemy. Changed the way in which Game.GameOver() is called. Now, Player.Die() is simply for assigning to the Player.IsDead property and the Game.GameOver call actually occurs after drawing everything in the game loop, to make the process more visually apparent to the player. Added an overload of Map.GetNeighbouringTiles() and altered both overloads to ommit any diagonal tiles from the returned arrays.