Added Items Namespace
Added Item class as a base for all items which can be picked up and used by the player. Added Inventory class as a container of items for the player and other entities. Removed Debug.Warning() call from Pathfinding class. Added ToFloat() method to Conversion class. Added MeleeWeapon class so that the player can pickup a weapon and increase their strength. Added Pickup class to act as a container entity which will hold items that can be picked up. Added a configuration menu to Placement.Place() so that pickups can be edited on placement. Added functionality to Entity.OnDestroy(), so that entities do not remain associated with a tile after being destroyed. Added pickup to Names dictionary in EntityCollection class. Implemented inventory functionality to Player class, so that items can be added, removed and equipped. Added extra validation to Game.Start(), so that if a player is not found in a map, the program does not crash. Added Input.PlayerInventory() so that the player can view and edit their inventory in-game. Added pickup case to MapFormatter.Deserialize().