Updated FileHandler & Fixed Bugs
Updated FileHandler so that the file structure is created in %AppData% on Windows and Application Support on macOS, instead of placing the files locally, which will not work for a published build of the game. Fixed a bug in the Placement class where cancelling the item configuration screen would crash the program. Fixed a bug in Player.EquippedItem where the player's strength would not return to two after unequipping a weapon or equipping a different item.