Fixed MapEditor Debugging Bug & Food Consuming Bug
Fixed a display bug in the editor where the contents of any container would be shown in the debugging menu in the map editor. Now only the type ID, entity ID and position are shown. Fixed a bug where food can be consumed even if the player has full health. This is now no longer the case.