Added Database Connection Validation
Altered GUI menu for incorrect log in details. Altered DatabaseManager.IsConnected property to check if the connection is not null. Added validation to DatabaseManager.Connect(). Altered Display.Clear() by swapping lines. Added GUI.Confirm() method to pause the flow of the program. Implemented GUI.Confirm() in MapEditor. In Program class, moved database connection functionality to Program.ConnectToDatabase(). This method is now called at the beginning, and whenever the user chooses the account setting when they are not already connected.