Added MapBrowser Class & MapProfile Structure
Added MapBrowser class with rudimentary functionality, which simply shows a vertical list containing the respective names and authors of the maps in the Map table of the database. Added DatabaseManager.QuerySQLIntoTable() which returns the result of an SQL query into a DataTable, allowing easy access to data. Altered catch block of DatabaseManager.Connect() to display an error screen with a confirmation, making this behaviour in Program.Main() mostly redundant. Cleaned up some unnecessary and inconsistent lines in DatabaseManager. Added MapProfile structure which will be used in future commits for storing data about maps contained on the database. Added functionality to the 'Online Maps' choice in the main menu in Program.Main().