Added Basic MapBrowser Download Functionality
Added Download() method to MapBrowser class, allowing maps to be downloaded from the database. Also added Upload() method to MapBrowser which adds some basic functionality for uploading maps to the database, provided the currently logged in account is an author account, although this method will likely need to be altered in future commits. Added AuthorID property to AuthorAccount class. Removed bracket inconsistencies from DatabaseManager class. Refined file path in MapEditor.Save() so that an interpolated string is used rather than ordinary string concatenation. Added shorthand question mark notation in the constructor of the Player class so that the current game's camera position is only set if the current game is not equal to null.