Validated MapBrowser and AccountManager
Updated MapBrowser.ShowBrowserMenu() to show the respective number of downloads alongside each map. Added validation to MapBrowser.Upload() so that user accounts cannot upload maps. Also added validation in the form of the EntityCollection. ContainsEntityOfType() method to ensure that only maps containing a player entity can be uploaded. Added Downloads property to MapProfile structure and a new overload of the MapProfile constructor.