Added Email Validation
Added AccountManager.ValidateEmailAddress() to check whether an email address entered by a user is valid or not by performing a presence check on an '@' symbol within the email string and by then checking that there are characters on either side of the '@' symbol. Also removed unnecessary using statement from MapBrowser class.