Added Exit Class
Added Exit class, which derives from Entity, in order to allow players to finish a map. Fixed a bug in Scoreboard.IsHighScore() which now checks if any scores exist in the Scoreboard table of the database. Added new Exit entity to Placement class and the EntityCollection structure as well as the MapFormatter.Deserialize() method. Added Entity.Push() to allow entities to be 'pushed' or interacted with by the player.  Removed unnecessary using statements from Pickup class. Added Game.Finish() which congratulates players, then allows them to submit their score and leave the current game.