Added HashManager Class & Updated AccountManager & EncryptionManager Classes
Added HashManager class to allow for the hashing of passwords, so that passwords cannot be easily 'decrypted' when accessed from the database. Since the EncryptionManager class is now only used for email addresses, the Encrypt and Decrypt methods have been renamed and altered to allow for an '@' symbol delimiter. These changes have been applied to the AccountManager class.