Added Water & Cleaned Up Sourcecode Formatting
Added a water tile which causes entities to be drawn slightly differently – as if they are partially submerged in the water. Also made minor formatting changes throughout the sourcecode to maintain consistency. Also added comments to parts of the sourcecode which make the least sense on their own. Also added some missing method and class summaries.