Bug Fixes & General Improvements
Made the Game class safer, so that deprecated or corrupted maps do not cause the program to crash due to a NullReferenceException. Appended the string of illegalCharacters in the MapEditor class. Altered EntityCollection.Dispose() to call its Remove() method on each entity as it is disposed of. Altered Player.Load() so that the EquippedItem property is correctly assigned to and does not crash the program if the player has no inventory items. Forgot to change the help keybind in the Input class from P to 0 in the last commit, so that has now been fixed. Altered colours of grass tile slightly in TileSet class. Added question mark notation to references to the Game.Current property in the Program and MapBrowser classes. Added Game class validation to the SaveManager class.